listservice/requirements.txt
django>=2.0,<3.0
psycopg2-binary>=2.7,<3.0
requests
gunicorn
djangorestframework>=3.10.3,<4.0
